Fixed problems with asynPrint formats
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15796 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/ADSrc/asynNDArrayDriver.cpp
@@ -375,8 +375,8 @@ asynStatus asynNDArrayDriver::readGenericPointer(asynUser *pasynUser, void *gene
     }
     if (!status)
         asynPrint(pasynUser, ASYN_TRACEIO_DRIVER,
-              "%s:%s: error, maxBytes=%d, data=%p\n",
-              driverName, functionName, arrayInfo.totalBytes, pArray->pData);
+              "%s:%s: error, maxBytes=%lu, data=%p\n",
+              driverName, functionName, (unsigned long)arrayInfo.totalBytes, pArray->pData);
     this->unlock();
     return status;
 }
